+++ cli/cli.go
@@ -1,4 +1,4 @@
-package main
+package cli
 
 import (
 	"fmt"
@@ -11,6 +11,7 @@ import (
 	"github.com/k0kubun/pp"
 
 	"github.com/creasty/rid/docker"
+	"github.com/creasty/rid/util"
 )
 
 const helpTemplate = `Execute commands via docker-compose
@@ -185,7 +186,7 @@ func (c *CLI) ExecHelp() error {
 		if cmd.HelpFile == "" {
 			continue
 		}
-		cmd.Summary, _ = loadHelpFile(cmd.HelpFile)
+		cmd.Summary, _ = util.LoadHelpFile(cmd.HelpFile)
 	}
 
 	tmpl := template.Must(template.New("help").Parse(helpTemplate))
@@ -198,7 +199,7 @@ func (c *CLI) ExecHelp() error {
 
 // ExecSubHelp shows help contents for a custom sub-command
 func (c *CLI) ExecSubHelp() error {
-	_, description := loadHelpFile(c.Args[1])
+	_, description := util.LoadHelpFile(c.Args[1])
 	fmt.Fprint(c.Stderr, description)
 	return nil
 }
