+++ main.go
@@ -4,6 +4,8 @@ import (
 	"fmt"
 	"os"
 	"os/exec"
+
+	"github.com/creasty/rid/cli"
 )
 
 func main() {
@@ -12,18 +14,18 @@ func main() {
 		exit(err)
 	}
 
-	ctx, err := NewContext(wd)
+	ctx, err := cli.NewContext(wd)
 	if err != nil {
 		exit(err)
 	}
 
-	cfg, err := NewConfig(ctx.ConfigFile)
+	cfg, err := cli.NewConfig(ctx.ConfigFile)
 	if err != nil {
 		exit(err)
 	}
 
-	cli := NewCLI(ctx, cfg, os.Args)
-	if err := cli.Run(); err != nil {
+	c := cli.NewCLI(ctx, cfg, os.Args)
+	if err := c.Run(); err != nil {
 		exit(err)
 	}
 }
